\name{spuRs-package}
\alias{spuRs-package}
\alias{spuRs}
\docType{package}
\title{
  "Introduction to Scientific Programming and Simulation Using R."
}
\description{
   This package provides scripts, functions and datasets for the book
  "Introduction to Scientific Programming and Simulation Using R."   
}
\details{
\tabular{ll}{
Package: \tab spuRs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-08-12\cr
License: \tab TBA\cr
LazyLoad: \tab yes\cr
}
  The datasets are: kew, treeg, trees, ufc.plots, and ufc.

  The main functions are: bisection, fixedpoint,
  fixedpoint\_show, newtonraphson, newtonraphson\_show, prime,
  primesieve, and vol.m3.

  We also provide functions relevant to the seedtrap case study: mean
  and print functions, an sd function, simulate, fitDistances, and
  constructors.  See Chapters 8 and 21 for more details.

  Type ?<object> to learn more about these objects, e.g. ?ufc
}
\author{
Owen Jones, Robert Maillardet, Andrew Robinson, Olga Borovkova, and Steve
Carnie.
Maintainer: Andrew Robinson <A.Robinson@ms.unimelb.edu.au>
}
\references{
  Jones, O.D., R.J. Maillardet, and A.P. Robinson.  2009.  Introduction
  to Scientific Programming and Simulation Using R.  Chapman And Hall/CRC.
}
\keyword{ package }
