\name{kew}
\alias{kew}
\docType{data}
\title{ 303 years of monthly rainfall data from Kew Gardens, London, U.K.}
\description{
  The monthly rainfall at Kew Gardens, London, U.K., from 1697 to 1999, in
  mm.
}
\usage{data(kew)}
\format{
  A wide-format data frame with 303 observations.  Each month has its
  own column.
}
\source{
  Data obtained from the U.S. National Climatic Data Centre, Global
  Historical Climatology Network data base (GHCN-Monthly Version 2, NB:
  not Version 3)
  \url{https://www.ncdc.noaa.gov/ghcnm/}.
}
\references{
  Jones, O.D., R. Maillardet, and A.P. Robinson.  2014.  An Introduction
  to Scientific Programming and Simulation, Using R.  2nd Ed. Chapman And Hall/CRC.
}
\examples{
data(kew)
}
\keyword{datasets}
